Elasticsearch #17.

+++ database/db-course.js
@@ -561,23 +561,26 @@ const DbCourse = class DbCourse extends DbObject {
     }
 
     async _updateSearchIndex(id, affected_list) {
-        let _affected = affected_list ? affected_list : [];
-        let result = ElasticConWrapper(async conn => {
-            let { deleted } = await IdxCourseService().importData(conn, { id: id, deleteIfNotExists: true, refresh: "true" });
-            for (let i = 0; i < _affected.length; i++) {
-                let service = _affected[i];
-                switch (service) {
-                    case "lesson":
-                        if (deleted > 0)
-                            await IdxLessonService().delete(conn, { courseId: id, refresh: "true" })
-                        else
-                            await IdxLessonService().importData(conn, { courseId: id, page: 5, refresh: "true" });
-                        break;
-                    default:
-                        throw new Error(`DbCourse::_updateSearchIndex: Unknown service name: "${service}".`);
+        let result = null;
+        if (config.has('search.keep_up_to_date') && config.get('search.keep_up_to_date')) {
+            let _affected = affected_list ? affected_list : [];
+            result = ElasticConWrapper(async conn => {
+                let { deleted } = await IdxCourseService().importData(conn, { id: id, deleteIfNotExists: true, refresh: "true" });
+                for (let i = 0; i < _affected.length; i++) {
+                    let service = _affected[i];
+                    switch (service) {
+                        case "lesson":
+                            if (deleted > 0)
+                                await IdxLessonService().delete(conn, { courseId: id, refresh: "true" })
+                            else
+                                await IdxLessonService().importData(conn, { courseId: id, page: 5, refresh: "true" });
+                            break;
+                        default:
+                            throw new Error(`DbCourse::_updateSearchIndex: Unknown service name: "${service}".`);
+                    }
                 }
-            }
-        }, true);
+            }, true);
+        }
         return result;
     }
